// frontend/src/shared/api/generated/employers/models/companyIn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Employers
 * OpenAPI spec version: 0.1.0
 */
import type { CompanySize } from './companySize';
import type { Industry } from './industry';

export interface CompanyIn {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  legal_name?: string | null;
  inn?: string | null;
  industry: Industry;
  /**
     * @minLength 1
     * @maxLength 8000
     */
  description: string;
  website?: string | null;
  city?: string | null;
  size?: CompanySize | null;
  /**
     * @maxItems 50
     * @items.minLength 1
     * @items.maxLength 64
     */
  tech_stack?: string[];
  contact_name?: string | null;
  contact_email?: string | null;
  contact_phone?: string | null;
  telegram?: string | null;
}